actually move Cursor struct to Model
