+++ src/model/mod.rs
@@ -1,3 +1,4 @@
 pub mod app;
 pub mod buffer;
+pub mod cursor;
 mod nodelist;
