+++ src/utils/mod.rs
@@ -1,4 +1,3 @@
-pub mod cursor;
 pub mod events;
 
 pub enum QuitOption {
